+++ src/tests/Main.Test.cpp
@@ -29,6 +29,8 @@ void registerTestAuxDataTypes()
     gtirb::AuxDataContainer::registerAuxDataType<LibraryPaths>();
     gtirb::AuxDataContainer::registerAuxDataType<SymbolicExpressionSizes>();
     gtirb::AuxDataContainer::registerAuxDataType<DdisasmVersion>();
+    gtirb::AuxDataContainer::registerAuxDataType<ElfStackSize>();
+    gtirb::AuxDataContainer::registerAuxDataType<ElfStackExec>();
 }
 
 int main(int argc, char** argv)